+++ libwasi/wasi.c
@@ -1508,128 +1508,6 @@ fail:
 }
 
 int
-wasi_clock_res_get(struct exec_context *ctx, struct host_instance *hi,
-                   const struct functype *ft, const struct cell *params,
-                   struct cell *results)
-{
-        WASI_TRACE;
-        HOST_FUNC_CONVERT_PARAMS(ft, params);
-        uint32_t clockid = HOST_FUNC_PARAM(ft, params, 0, i32);
-        uint32_t retp = HOST_FUNC_PARAM(ft, params, 1, i32);
-        clockid_t hostclockid;
-        int host_ret = 0;
-        int ret;
-        ret = wasi_convert_clockid(clockid, &hostclockid);
-        if (ret != 0) {
-                goto fail;
-        }
-        struct timespec ts;
-        ret = clock_getres(hostclockid, &ts);
-        if (ret == -1) {
-                ret = errno;
-                assert(ret > 0);
-                goto fail;
-        }
-        uint64_t result = host_to_le64(timespec_to_ns(&ts));
-        host_ret = wasi_copyout(ctx, &result, retp, sizeof(result),
-                                WASI_U64_ALIGN);
-fail:
-        if (host_ret == 0) {
-                HOST_FUNC_RESULT_SET(ft, results, 0, i32,
-                                     wasi_convert_errno(ret));
-        }
-        HOST_FUNC_FREE_CONVERTED_PARAMS();
-        return host_ret;
-}
-
-int
-wasi_clock_time_get(struct exec_context *ctx, struct host_instance *hi,
-                    const struct functype *ft, const struct cell *params,
-                    struct cell *results)
-{
-        WASI_TRACE;
-        HOST_FUNC_CONVERT_PARAMS(ft, params);
-        uint32_t clockid = HOST_FUNC_PARAM(ft, params, 0, i32);
-#if 0 /* REVISIT what to do with the precision? */
-        uint64_t precision = HOST_FUNC_PARAM(ft, params, 1, i64);
-#endif
-        uint32_t retp = HOST_FUNC_PARAM(ft, params, 2, i32);
-        clockid_t hostclockid;
-        int host_ret = 0;
-        int ret;
-        ret = wasi_convert_clockid(clockid, &hostclockid);
-        if (ret != 0) {
-                goto fail;
-        }
-        struct timespec ts;
-        ret = clock_gettime(hostclockid, &ts);
-        if (ret == -1) {
-                ret = errno;
-                assert(ret > 0);
-                goto fail;
-        }
-        uint64_t result = host_to_le64(timespec_to_ns(&ts));
-        host_ret = wasi_copyout(ctx, &result, retp, sizeof(result),
-                                WASI_U64_ALIGN);
-fail:
-        if (host_ret == 0) {
-                HOST_FUNC_RESULT_SET(ft, results, 0, i32,
-                                     wasi_convert_errno(ret));
-        }
-        HOST_FUNC_FREE_CONVERTED_PARAMS();
-        return host_ret;
-}
-
-int
-wasi_random_get(struct exec_context *ctx, struct host_instance *hi,
-                const struct functype *ft, const struct cell *params,
-                struct cell *results)
-{
-        WASI_TRACE;
-        HOST_FUNC_CONVERT_PARAMS(ft, params);
-        uint32_t buf = HOST_FUNC_PARAM(ft, params, 0, i32);
-        uint32_t buflen = HOST_FUNC_PARAM(ft, params, 1, i32);
-        int ret = 0;
-        void *p;
-        int host_ret = memory_getptr(ctx, 0, buf, 0, buflen, &p);
-        if (host_ret != 0) {
-                goto fail;
-        }
-#if defined(__GLIBC__) || defined(__NuttX__)
-        /*
-         * glibc doesn't have arc4random
-         * https://sourceware.org/bugzilla/show_bug.cgi?id=4417
-         *
-         * NuttX has both of getrandom and arc4random_buf.
-         * The latter is available only if CONFIG_CRYPTO_RANDOM_POOL=y.
-         */
-        while (buflen > 0) {
-                ssize_t ssz = getrandom(p, buflen, 0);
-                if (ssz == -1) {
-                        ret = errno;
-                        assert(ret > 0);
-                        if (ret == EINTR) {
-                                continue;
-                        }
-                        break;
-                }
-                p = (uint8_t *)p + ssz;
-                buflen -= ssz;
-        }
-#else
-        arc4random_buf(p, buflen);
-        ret = 0;
-#endif
-fail:
-        if (host_ret == 0) {
-                HOST_FUNC_RESULT_SET(ft, results, 0, i32,
-                                     wasi_convert_errno(ret));
-        }
-        HOST_FUNC_FREE_CONVERTED_PARAMS();
-        return host_ret;
-}
-
-int
 wasi_path_open(struct exec_context *ctx, struct host_instance *hi,
                const struct functype *ft, const struct cell *params,
                struct cell *results)
